1. js-code was splitted by modules
2. function getRandomNumber was refactored
